requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.29.0
python-dotenv==1.0.1
pydantic==2.5.3
requests==2.31.0
pandas==2.0.3
mlflow==2.8.0
torch==2.1.0
transformers==4.36.0
datasets==2.16.0
peft==0.7.0
accelerate==0.26.0